+++ scrollpointclick.AI/src/index.js
@@ -22,7 +22,6 @@
 
 export {
   testConnection,
-  createAIImages,
   createResearchRequest,
   createResearchListRequest,
   createQuickSearch,
@@ -34,6 +33,7 @@ export {
 } from './NPAI' // add one of these for every command specifified in plugin.json (the function could be in any file as long as it's exported)
 export { bulletsAI, createResearchDigSite, remixQuery, explore, researchFromSelection } from './BulletsAI-Main'
 export { adjustPreferences, scrollToEntry, retrieveResearchNotes } from './support/helpers'
+export { createAIImages } from './imageAI'
 
 // import './support/fetchOverrides'